Stage 5

+++ src/main/java/com/diarpy/recipes/businessLayer/RecipeService.java
@@ -2,30 +2,30 @@ package com.diarpy.recipes.businessLayer;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 import com.diarpy.recipes.persistance.RecipeRepository;
+import com.diarpy.recipes.persistance.UserRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 
-/**
- * @author Mack_TB
- * @version 1.0.7
- * @since 9/9/2021
- */
-
 @Service
 public class RecipeService {
     private final RecipeRepository recipeRepository;
+    private final UserRepository userRepository;
 
     @Autowired
-    public RecipeService(RecipeRepository recipeRepository) {
+    public RecipeService(RecipeRepository recipeRepository, UserRepository userRepository) {
         this.recipeRepository = recipeRepository;
+        this.userRepository = userRepository;
     }
 
-    public Recipe save(Recipe newRecipe) {
+    public Recipe save(Recipe newRecipe, Authentication authentication) {
+        User user = userRepository.findByEmail(authentication.getName());
+        newRecipe.setUser(user);
         return recipeRepository.save(newRecipe);
     }
 
@@ -36,14 +36,23 @@ public class RecipeService {
                                 "Recipe not found for id = " + id));
     }
 
-    public void deleteRecipeById(Long id) {
+    public void deleteRecipeById(Long id, Authentication authentication) {
         Recipe recipe = findRecipeById(id);
+        User user = userRepository.findByEmail(authentication.getName());
+        if (recipe.getUser() != user) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+                    "You are not allowed");
+        }
         recipeRepository.delete(recipe);
     }
 
-    public void updateRecipe(Long id, Recipe recipe) {
+    public void updateRecipe(Long id, Recipe recipe, Authentication authentication) {
         Recipe storedRecipe = findRecipeById(id);
-
+        User user = userRepository.findByEmail(authentication.getName());
+        if (storedRecipe.getUser() != user) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN,
+                    "You are not allowed");
+        }
         storedRecipe.setName(recipe.getName());
         storedRecipe.setCategory(recipe.getCategory());
         storedRecipe.setDescription(recipe.getDescription());
@@ -72,5 +81,4 @@ public class RecipeService {
         }
         return recipes;
     }
-
 }
